FTR/FIX: Web interface for login functional. Need to create a way for the search interface to access the API rank call.
