+++ pydatarecognition/dependencies.py
@@ -1,0 +1,14 @@
+from typing import Optional
+from starlette.requests import Request
+from fastapi import HTTPException
+
+
+# Try to get the logged in user
+async def get_user(request: Request) -> Optional[dict]:
+    user = request.session.get('user')
+    if user is not None:
+        return user
+    else:
+        raise HTTPException(status_code=403, detail='Could not validate credentials.')
+
+    return None
