+++ pydatarecognition/templates/footer-about.html
@@ -6,7 +6,7 @@
 
     <div class="w3-cell-row" style="position:relative;text-align:middle;margin: 0 auto;float: none;max-width: 1100px">
         <div class="w3-section">
-            <h3>About the PDF in the Cloud</h3>
+            <h3>About the CIF Powder Search</h3>
             <p>
                 This is a cloud-based platform for CIF based literature search.
             </p>
@@ -36,25 +36,25 @@
             </p>
 
             <h4 style="padding-top: 5px">Co-Investigators</h4>
-<!--            <p>-->
-<!--                <a href="https://www.ornl.gov/staff-profile/matthew-g-tucker" style="color:blue;text-decoration: none" target="_blank">Matt Tucker </a> Oak Ridge National Laboratory-->
-<!--            </p>-->
-<!--            <p>-->
-<!--                <a href="https://chem.ku.dk/ansatte/alle/?pure=en/persons/540779" style="color:blue;text-decoration: none" target="_blank">Kirsten Jensen </a> University of Copenhagen-->
-<!--            </p>-->
+{#            <p>#}
+{#                <a href="https://www.ornl.gov/staff-profile/matthew-g-tucker" style="color:blue;text-decoration: none" target="_blank">Matt Tucker </a> Oak Ridge National Laboratory#}
+{#            </p>#}
+{#            <p>#}
+{#                <a href="https://chem.ku.dk/ansatte/alle/?pure=en/persons/540779" style="color:blue;text-decoration: none" target="_blank">Kirsten Jensen </a> University of Copenhagen#}
+{#          </p>#}
         </div>
 
         <div class="w3-section" style="padding-top: 10px;">
             <h3>Partners and Support</h3>
           <img src="/static/img/columbia_bw.png" width="100px">
-<!--          <img src="/static/img/nsf_bw.png" width="100px">-->
-<!--            <p>-->
-<!--                This material is based upon work supported by the National Science Foundation under Grant DMREF-1534910.-->
-<!--            </p>-->
-<!--            <p>-->
-<!--                Any opinions, findings, and conclusions or recommendations expressed in this material-->
-<!--                are those of the author(s) and do not necessarily reflect the views of the National Science Foundation.-->
-<!--            </p>-->
+{#          <img src="/static/img/nsf_bw.png" width="100px">#}
+{#            <p>#}
+{#                This material is based upon work supported by the National Science Foundation under Grant DMREF-1534910.#}
+{#            </p>#}
+{#            <p>#}
+{#                Any opinions, findings, and conclusions or recommendations expressed in this material#}
+{#              are those of the author(s) and do not necessarily reflect the views of the National Science Foundation.#}
+{#          </p>#}
 
         </div>
