+++ pydatarecognition/templates/footer-privacy.html
@@ -9,7 +9,8 @@
         </div>
         <div class="w3-section">
             <p>
-<!--                Use of this web site is subject to the <a href="{{url_for('static', filename='privacy_policy/privacy_policy_2021-07-05.pdf') }}" style="color:blue;text-decoration: none" target="_blank">Privacy Policy</a> described in the <a href="{{url_for('static', filename='privacy_policy/privacy_policy_2021-07-05.pdf') }}" style="color:blue;text-decoration: none" target="_blank">attached document</a>.-->
+                Please be very private...
+{#                Use of this web site is subject to the <a href="{{url_for('static', path='/privacy_policy/privacy_policy_2021-07-05.pdf') }}" style="color:blue;text-decoration: none" target="_blank">Privacy Policy</a> described in the <a href="{{url_for('static', filename='privacy_policy/privacy_policy_2021-07-05.pdf') }}" style="color:blue;text-decoration: none" target="_blank">attached document</a>.#}
             </p>
         </div>
     </div>
