+++ pydatarecognition/templates/footer-term.html
@@ -9,7 +9,8 @@
         </div>
         <div class="w3-section">
             <p>
-<!--                Use of this web site is subject to the <a href="{{url_for('static', filename='terms_of_use/terms_of_use_2021-07-05.pdf') }}" style="color:blue;text-decoration: none" target="_blank">Terms of Use</a> described in the <a href="{{url_for('static', filename='terms_of_use/terms_of_use_2021-07-05.pdf') }}" style="color:blue;text-decoration: none" target="_blank">attached document</a>.-->
+                Please stay on good terms...
+{#                Use of this web site is subject to the <a href="{{url_for('static', path='/terms_of_use/terms_of_use_2021-07-05.pdf') }}" style="color:blue;text-decoration: none" target="_blank">Terms of Use</a> described in the <a href="{{url_for('static', filename='terms_of_use/terms_of_use_2021-07-05.pdf') }}" style="color:blue;text-decoration: none" target="_blank">attached document</a>.#}
             </p>
 
         </div>
